+++ superuser_plugin_unix/lib/superuser_plugin_unix.dart
@@ -12,37 +12,47 @@ const String _libName = 'superuser_plugin_unix';
 ///
 /// Remark: [isActivated] is identical with [isSuperuser]
 /// since `root` is a definition of superuser.
-final class UnixSuperuser implements SuperuserInterface {
-  static UnixSuperuser? _instance;
+final class UnixSuperuser extends SuperuserPlatform {
+  UnixSuperuser()
+      : super(() {
+          if (Platform.isMacOS) {
+            return DynamicLibrary.open('$_libName.framework/$_libName');
+          }
 
-  late final SuperuserPluginUnixBindings _bindings;
+          if (Platform.isLinux) {
+            return DynamicLibrary.open('lib$_libName.so');
+          }
 
-  UnixSuperuser._() {
-    _bindings = SuperuserPluginUnixBindings(() {
-      if (Platform.isMacOS) {
-        return DynamicLibrary.open('$_libName.framework/$_libName');
-      }
-
-      if (Platform.isLinux) {
-        return DynamicLibrary.open('lib$_libName.so');
-      }
-
-      throw UnsupportedError('Unknown platform: ${Platform.operatingSystem}');
-    }());
-  }
-
-  factory UnixSuperuser() {
-    _instance ??= UnixSuperuser._();
-
-    return _instance!;
-  }
+          throw UnsupportedError(
+              'Unknown platform: ${Platform.operatingSystem}');
+        });
 
   @override
   bool get isActivated => isSuperuser;
 
   @override
-  bool get isSuperuser => _bindings.is_root();
+  bool get isSuperuser =>
+      onGettingProperties((lib) => SuperuserPluginUnixBindings(lib).is_root());
 
   @override
-  String get whoAmI => _bindings.get_uname().cast<ffi.Utf8>().toDartString();
+  String get whoAmI => onGettingProperties((lib) {
+        final SuperuserPluginUnixBindings bindings =
+            SuperuserPluginUnixBindings(lib);
+
+        Pointer<Pointer<Char>> resultPtr = ffi.calloc<Pointer<Char>>();
+
+        try {
+          int errCode = bindings.get_uname(resultPtr);
+
+          if (errCode != 0) {
+            throw SuperuserProcessError(errCode, "Unable to retrive username.");
+          }
+
+          String result = resultPtr.value.cast<ffi.Utf8>().toDartString();
+
+          return result;
+        } finally {
+          ffi.calloc.free(resultPtr);
+        }
+      });
 }