+++ superuser_plugin_unix/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0
+
+* Add throw error if native code returns with error.
+
 ## 0.1.0
 
 * Define interface's result