+++ superuser_plugin_windows/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.0
+
+* Add throw error if native code returns with error.
+* Improve memory effiency.
+
 ## 0.1.0
 
 * Define interface's result